Create HashSet.java
Problem Description
Create a HashSet implementation without relying on any pre-existing hash table libraries.

You are tasked with implementing the following functions:

void add(key): Inserts the value key into the HashSet.
bool contains(key): Indicates whether the value key exists in the HashSet.
void remove(key): Deletes the value key from the HashSet. If key does not exist in the HashSet, no action is taken.
Input Format: 

You are given a sequence of Q queries, each designated by a value ranging from 1 to 3.

For each query, a key is provided as well.

The query values signify the following:

add(): Call to the add() function with a key as a parameter.
contains(): Call to the contains() function with a key as a parameter.
remove(): Call to the remove() function with a key as a parameter.